+++ DREAM Lab Frontend/src/Home/RecommendationsCarousel.css
@@ -1,3 +1,5 @@
+@import url('https://fonts.googleapis.com/css2?family=Karla:ital,wght@0,200..800;1,200..800&display=swap');
+
 .imageCarousel {
 	user-select: none;
 	width: 100%;
@@ -11,6 +13,10 @@
     justify-content: center;
 }
 
-.transcriptOverlay {
+.processed-transcript-container {
+	display: flex;
+    justify-content: center;
+	font-family: 'Karla', sans-serif;
 	color: white;
+	font-size: large;
 }
